Added some player logic
Nothing should have visibly changed, but now the game will look up the
player object associated with your login name on the website, and then
look up the location of the player object to get to the location object.

It also defaults the player object to "Kazetsukai" if it can't find one
lol.