Muhahaha! Sweeping destruction!
I have removed all the MVC stuff. We now have a project that is a
Self-Hosted WebApi server, with data APIs and a single-page-app style
index.html being served up. Angular has been added, and is working, but
the project itself is not functional yet. Feel free to get the location
descriptions going again in angular if you beat me to it, and we will
need a good starter CSS in there I think.